+++ invoice_test.go
@@ -236,76 +236,6 @@ func TestCustomer(t *testing.T) {
 	assert.Equal(t, "me.unselfish@me.com", gi.Customer.Emails[0].Address)
 }
 
-func TestTaxInclusive(t *testing.T) {
-
-	s := validStripeInvoice()
-
-	// Check that tax are inclusive
-	gi, err := goblstripe.FromInvoice(s, uuid.MustParse(namespace))
-	require.NoError(t, err)
-	assert.Nil(t, gi.Tax)
-
-	s.TotalTaxAmounts[0].Inclusive = true
-	gi, err = goblstripe.FromInvoice(s, uuid.MustParse(namespace))
-	require.NoError(t, err)
-	assert.Equal(t, tax.CategoryVAT, gi.Tax.PricesInclude)
-}
-
-func TestOrderingPeriod(t *testing.T) {
-	s := validStripeInvoice()
-	gi, err := goblstripe.FromInvoice(s, uuid.MustParse(namespace))
-	require.NoError(t, err)
-
-	assert.Equal(t, "2025-01-08", gi.Lines[0].Item.Meta[goblstripe.MetaKeyDateFrom])
-	assert.Equal(t, "2025-02-08", gi.Lines[0].Item.Meta[goblstripe.MetaKeyDateTo])
-}
-
-func TestShippingDetails(t *testing.T) {
-	s := validStripeInvoice()
-	gi, err := goblstripe.FromInvoice(s, uuid.MustParse(namespace))
-	require.NoError(t, err)
-
-	assert.Equal(t, "Test Customer", gi.Delivery.Receiver.Name)
-	assert.Equal(t, "Berlin", gi.Delivery.Receiver.Addresses[0].Locality)
-	assert.Equal(t, "DE", gi.Delivery.Receiver.Addresses[0].Country.String())
-	assert.Equal(t, "Unter den Linden 1", gi.Delivery.Receiver.Addresses[0].Street)
-	assert.Equal(t, "10117", gi.Delivery.Receiver.Addresses[0].Code.String())
-	assert.Equal(t, "BE", gi.Delivery.Receiver.Addresses[0].State.String())
-}
-
-func TestNoTermsWhenPaid(t *testing.T) {
-	s := validStripeInvoice()
-	gi, err := goblstripe.FromInvoice(s, uuid.MustParse(namespace))
-	require.NoError(t, err)
-
-	assert.Nil(t, gi.Payment)
-}
-
-func TestExchangeRateConversionDefault(t *testing.T) {
-	s := validStripeInvoice()
-	gi, err := goblstripe.FromInvoice(s, uuid.MustParse(namespace))
-	require.NoError(t, err)
-
-	assert.Nil(t, gi.ExchangeRates)
-
-	s.Currency = "usd"
-	gi, err = goblstripe.FromInvoice(s, uuid.MustParse(namespace))
-	require.NoError(t, err)
-
-	assert.Equal(t, num.MakeAmount(935, 3), gi.ExchangeRates[0].Amount)
-}
-
-func TestZeroDecimalCurrencies(t *testing.T) {
-	s := validStripeInvoice()
-	s.Currency = "jpy"
-	s.Lines.Data[0].Currency = "jpy"
-	gi, err := goblstripe.FromInvoice(s, uuid.MustParse(namespace))
-	require.NoError(t, err)
-
-	assert.Equal(t, currency.JPY, gi.Currency)
-	assert.Equal(t, num.MakeAmount(-11000, 0), gi.Lines[0].Item.Price)
-}
-
 func TestCalculate(t *testing.T) {
 	s := validStripeInvoice()
 	gi, err := goblstripe.FromInvoice(s, uuid.MustParse(namespace))
