+++ party.go
@@ -16,6 +16,86 @@ import (
 
 var orgIDKeys = []stripe.TaxIDType{"de_stn"}
 
+// Lookup map for country code to Stripe tax ID type
+var taxIDMapGOBLToStripe = map[l10n.Code]stripe.TaxIDType{
+	l10n.AL: stripe.TaxIDTypeAlTin,
+	l10n.AD: stripe.TaxIDTypeADNRT,
+	l10n.AO: stripe.TaxIDTypeAoTin,
+	l10n.AR: stripe.TaxIDTypeARCUIT,
+	l10n.AM: stripe.TaxIDTypeAmTin,
+	// TODO: handle other Australian tax ID type
+	l10n.AU: stripe.TaxIDTypeAUABN, // main code for businesses
+	l10n.BS: stripe.TaxIDTypeBsTin,
+	l10n.BH: stripe.TaxIDTypeBhVAT,
+	l10n.BB: stripe.TaxIDTypeBbTin,
+	l10n.BY: stripe.TaxIDTypeByTin,
+	l10n.BO: stripe.TaxIDTypeBOTIN,
+	l10n.BA: stripe.TaxIDTypeBaTin,
+	l10n.BR: stripe.TaxIDTypeBRCNPJ,
+	l10n.KH: stripe.TaxIDTypeKhTin,
+	l10n.CA: stripe.TaxIDTypeCABN, // TODO: correctly handle GST, HST, and PST variants
+	l10n.CL: stripe.TaxIDTypeCLTIN,
+	l10n.CN: stripe.TaxIDTypeCNTIN,
+	l10n.CO: stripe.TaxIDTypeCONIT,
+	l10n.CD: stripe.TaxIDTypeCdNif,
+	l10n.CR: stripe.TaxIDTypeCRTIN,
+	l10n.DO: stripe.TaxIDTypeDORCN,
+	l10n.EC: stripe.TaxIDTypeECRUC,
+	l10n.EG: stripe.TaxIDTypeEGTIN,
+	l10n.SV: stripe.TaxIDTypeSVNIT,
+	l10n.GE: stripe.TaxIDTypeGEVAT,
+	l10n.GN: stripe.TaxIDTypeGnNif,
+	l10n.HK: stripe.TaxIDTypeHKBR,
+	l10n.IS: stripe.TaxIDTypeISVAT,
+	l10n.IN: stripe.TaxIDTypeINGST,
+	l10n.ID: stripe.TaxIDTypeIDNPWP,
+	l10n.IL: stripe.TaxIDTypeILVAT,
+	l10n.JP: stripe.TaxIDTypeJPCN, // TODO: handle other Japanese tax ID types
+	l10n.KZ: stripe.TaxIDTypeKzBin,
+	l10n.KE: stripe.TaxIDTypeKEPIN,
+	l10n.LI: stripe.TaxIDTypeLiVAT, // TODO handle Liechtenstein UID number
+	l10n.MY: stripe.TaxIDTypeMYITN, // TODO handle Malaysian FRP and SST numbers
+	l10n.MR: stripe.TaxIDTypeMrNif,
+	l10n.MX: stripe.TaxIDTypeMXRFC,
+	l10n.MD: stripe.TaxIDTypeMdVAT,
+	l10n.ME: stripe.TaxIDTypeMePib,
+	l10n.MA: stripe.TaxIDTypeMaVAT,
+	l10n.NP: stripe.TaxIDTypeNpPan,
+	l10n.NZ: stripe.TaxIDTypeNZGST,
+	l10n.NG: stripe.TaxIDTypeNgTin,
+	l10n.MK: stripe.TaxIDTypeMkVAT,
+	l10n.OM: stripe.TaxIDTypeOmVAT,
+	l10n.PE: stripe.TaxIDTypePERUC,
+	l10n.PH: stripe.TaxIDTypePHTIN,
+	l10n.RO: stripe.TaxIDTypeROTIN,
+	l10n.RU: stripe.TaxIDTypeRUINN, // TODO handle KPP
+	l10n.SA: stripe.TaxIDTypeSAVAT,
+	l10n.SN: stripe.TaxIDTypeSnNinea,
+	l10n.RS: stripe.TaxIDTypeRSPIB,
+	l10n.SG: stripe.TaxIDTypeSGGST, // TODO: handle UEN
+	l10n.SI: stripe.TaxIDTypeSITIN,
+	l10n.ZA: stripe.TaxIDTypeZAVAT,
+	l10n.KR: stripe.TaxIDTypeKRBRN,
+	l10n.SR: stripe.TaxIDTypeSrFin,
+	l10n.CH: stripe.TaxIDTypeCHVAT, // TODO: handle UID
+	l10n.TW: stripe.TaxIDTypeTWVAT,
+	l10n.TJ: stripe.TaxIDTypeTjTin,
+	l10n.TZ: stripe.TaxIDTypeTzVAT,
+	l10n.TH: stripe.TaxIDTypeTHVAT,
+	l10n.TR: stripe.TaxIDTypeTRTIN,
+	l10n.UG: stripe.TaxIDTypeUgTin,
+	l10n.UA: stripe.TaxIDTypeUAVAT,
+	l10n.AE: stripe.TaxIDTypeAETRN,
+	l10n.GB: stripe.TaxIDTypeGBVAT,
+	l10n.US: stripe.TaxIDTypeUSEIN, // TODO: determine if this should be here
+	l10n.UY: stripe.TaxIDTypeUYRUC,
+	l10n.UZ: stripe.TaxIDTypeUzTin,
+	l10n.VE: stripe.TaxIDTypeVERIF,
+	l10n.VN: stripe.TaxIDTypeVNTIN,
+	l10n.ZM: stripe.TaxIDTypeZmTin,
+	l10n.ZW: stripe.TaxIDTypeZwTin,
+}
+
 // ToCustomerTaxIDDataParamsForTax takes a GOBL tax Identity and converts it into
 // the expected stripe tax ID object. GOBL tax identities are much more restrictive than
 // those used in Stripe and only accept a specific code type for the given country. Countries
@@ -32,185 +112,27 @@ func ToCustomerTaxIDDataParamsForTax(tID *tax.Identity) *stripe.CustomerTaxIDDat
 			Value: stripe.String(tID.String()), // Include the country code.
 		}
 	}
-	var typ stripe.TaxIDType
-	switch tID.Country.Code() {
-	case l10n.EU:
+
+	if tID.Country.Code() == l10n.EU {
 		return &stripe.CustomerTaxIDDataParams{
 			Type:  stripe.String(string(stripe.TaxIDTypeEUOSSVAT)),
 			Value: stripe.String(tID.String()), // Include the country code.
 		}
-	case l10n.AL:
-		typ = stripe.TaxIDTypeAlTin
-	case l10n.AD:
-		typ = stripe.TaxIDTypeADNRT
-	case l10n.AO:
-		typ = stripe.TaxIDTypeAoTin
-	case l10n.AR:
-		typ = stripe.TaxIDTypeARCUIT
-	case l10n.AM:
-		typ = stripe.TaxIDTypeAmTin
-	case l10n.AU:
-		// TODO: handle other Australian tax ID type
-		typ = stripe.TaxIDTypeAUABN // main code for businesses
-	case l10n.BS:
-		typ = stripe.TaxIDTypeBsTin
-	case l10n.BH:
-		typ = stripe.TaxIDTypeBhVAT
-	case l10n.BB:
-		typ = stripe.TaxIDTypeBbTin
-	case l10n.BY:
-		typ = stripe.TaxIDTypeByTin
-	case l10n.BO:
-		typ = stripe.TaxIDTypeBOTIN
-	case l10n.BA:
-		typ = stripe.TaxIDTypeBaTin
-	case l10n.BR:
-		typ = stripe.TaxIDTypeBRCNPJ
-	case l10n.KH:
-		typ = stripe.TaxIDTypeKhTin
-	case l10n.CA:
-		// TODO: correctly handle GST, HST, and PST variants
-		typ = stripe.TaxIDTypeCABN
-	case l10n.CL:
-		typ = stripe.TaxIDTypeCLTIN
-	case l10n.CN:
-		typ = stripe.TaxIDTypeCNTIN
-	case l10n.CO:
-		typ = stripe.TaxIDTypeCONIT
-	case l10n.CD:
-		typ = stripe.TaxIDTypeCdNif
-	case l10n.CR:
-		typ = stripe.TaxIDTypeCRTIN
-	case l10n.DO:
-		typ = stripe.TaxIDTypeDORCN
-	case l10n.EC:
-		typ = stripe.TaxIDTypeECRUC
-	case l10n.EG:
-		typ = stripe.TaxIDTypeEGTIN
-	case l10n.SV:
-		typ = stripe.TaxIDTypeSVNIT
-	case l10n.GE:
-		typ = stripe.TaxIDTypeGEVAT
-	case l10n.GN:
-		typ = stripe.TaxIDTypeGnNif
-	case l10n.HK:
-		typ = stripe.TaxIDTypeHKBR
-	case l10n.IS:
-		typ = stripe.TaxIDTypeISVAT
-	case l10n.IN:
-		typ = stripe.TaxIDTypeINGST
-	case l10n.ID:
-		typ = stripe.TaxIDTypeIDNPWP
-	case l10n.IL:
-		typ = stripe.TaxIDTypeILVAT
-	case l10n.JP:
-		// TODO handle other Japanese tax ID types
-		typ = stripe.TaxIDTypeJPCN
-	case l10n.KZ:
-		typ = stripe.TaxIDTypeKzBin
-	case l10n.KE:
-		typ = stripe.TaxIDTypeKEPIN
-	case l10n.LI:
-		// TODO handle Liechtenstein UID number
-		typ = stripe.TaxIDTypeLiVAT
-	case l10n.MY:
-		// TODO handle Malaysian FRP and SST numbers
-		typ = stripe.TaxIDTypeMYITN
-	case l10n.MR:
-		typ = stripe.TaxIDTypeMrNif
-	case l10n.MX:
-		typ = stripe.TaxIDTypeMXRFC
-	case l10n.MD:
-		typ = stripe.TaxIDTypeMdVAT
-	case l10n.ME:
-		typ = stripe.TaxIDTypeMePib
-	case l10n.MA:
-		typ = stripe.TaxIDTypeMaVAT
-	case l10n.NP:
-		typ = stripe.TaxIDTypeNpPan
-	case l10n.NZ:
-		typ = stripe.TaxIDTypeNZGST
-	case l10n.NG:
-		typ = stripe.TaxIDTypeNgTin
-	case l10n.MK:
-		typ = stripe.TaxIDTypeMkVAT
-	case l10n.OM:
-		typ = stripe.TaxIDTypeOmVAT
-	case l10n.PE:
-		typ = stripe.TaxIDTypePERUC
-	case l10n.PH:
-		typ = stripe.TaxIDTypePHTIN
-	case l10n.RO:
-		typ = stripe.TaxIDTypeROTIN
-	case l10n.RU:
-		// TODO handle KPP
-		typ = stripe.TaxIDTypeRUINN
-	case l10n.SA:
-		typ = stripe.TaxIDTypeSAVAT
-	case l10n.SN:
-		typ = stripe.TaxIDTypeSnNinea
-	case l10n.RS:
-		typ = stripe.TaxIDTypeRSPIB
-	case l10n.SG:
-		// TODO: handle UEN
-		typ = stripe.TaxIDTypeSGGST
-	case l10n.SI:
-		typ = stripe.TaxIDTypeSITIN
-	case l10n.ZA:
-		typ = stripe.TaxIDTypeZAVAT
-	case l10n.KR:
-		typ = stripe.TaxIDTypeKRBRN
-	case l10n.SR:
-		typ = stripe.TaxIDTypeSrFin
-	case l10n.CH:
-		// TODO: handle UID
-		typ = stripe.TaxIDTypeCHVAT
-	case l10n.TW:
-		typ = stripe.TaxIDTypeTWVAT
-	case l10n.TJ:
-		typ = stripe.TaxIDTypeTjTin
-	case l10n.TZ:
-		typ = stripe.TaxIDTypeTzVAT
-	case l10n.TH:
-		typ = stripe.TaxIDTypeTHVAT
-	case l10n.TR:
-		typ = stripe.TaxIDTypeTRTIN
-	case l10n.UG:
-		typ = stripe.TaxIDTypeUgTin
-	case l10n.UA:
-		typ = stripe.TaxIDTypeUAVAT
-	case l10n.AE:
-		typ = stripe.TaxIDTypeAETRN
-	case l10n.GB:
-		typ = stripe.TaxIDTypeGBVAT
-	case l10n.US:
-		// TODO: determine if this should be here
-		typ = stripe.TaxIDTypeUSEIN
-	case l10n.UY:
-		typ = stripe.TaxIDTypeUYRUC
-	case l10n.UZ:
-		typ = stripe.TaxIDTypeUzTin
-	case l10n.VE:
-		typ = stripe.TaxIDTypeVERIF
-	case l10n.VN:
-		typ = stripe.TaxIDTypeVNTIN
-	case l10n.ZM:
-		typ = stripe.TaxIDTypeZmTin
-	case l10n.ZW:
-		typ = stripe.TaxIDTypeZwTin
 	}
-	if typ != "" {
+
+	if typ, ok := taxIDMapGOBLToStripe[tID.Country.Code()]; ok {
 		return &stripe.CustomerTaxIDDataParams{
 			Type:  stripe.String(string(typ)),
 			Value: stripe.String(tID.Code.String()),
 		}
 	}
+
 	return nil
 }
 
-// ToTaxIDForOrg expects a GOBL org Identity and attempts to convert it into a
+// ToTaxIDFromOrg expects a GOBL org Identity and attempts to convert it into a
 // Stripe tax ID object.
-func ToTaxIDForOrg(id *org.Identity) *stripe.TaxID {
+func ToTaxIDFromOrg(id *org.Identity) *stripe.TaxID {
 	if id == nil {
 		return nil
 	}
@@ -265,7 +187,7 @@ func FromTaxIDToTax(taxID *stripe.TaxID) *tax.Identity {
 	return nil
 }
 
-// FromTaxIDForOrg converts a stripe tax ID object into a GOBL organization identity, if possible.
+// FromTaxIDToOrg converts a stripe tax ID object into a GOBL organization identity, if possible.
 func FromTaxIDToOrg(taxID *stripe.TaxID) *org.Identity {
 	if taxID == nil {
 		return nil
@@ -464,7 +386,7 @@ func FromTelephone(phone string) *org.Telephone {
 	}
 }
 
-// ToAddress converts the GOBL address object into a stripe address object.
+// ToAddressParams converts the GOBL address object into a stripe address object.
 func ToAddressParams(addr *org.Address) *stripe.AddressParams {
 	if addr == nil {
 		return nil
